update _parseData and saveFITS functions

+++ dslrpp/tools/stack.py
@@ -34,6 +34,6 @@ def stack(images, mode='median'):
         raise ValueError("Invalid argument for 'mode' parameter")
 
     return Monochrome(
-            stack.astype('uint8'), images[0], images[0].imcolor,
+            stack.astype('uint16'), images[0], images[0].imcolor,
             stacked=True
             )
